ingester/requirements.txt
redis>=5.0.0
python-dotenv
